Refactor worker interrupt handling into dedicated module
