+++ src/worker_process.rs
@@ -34,8 +34,7 @@ use crate::pending_output_tape::{FormattedPendingOutput, PendingOutputTape, Pend
 use crate::reply_presentation::trim_echo_prefix_after_leading_nonstdout_contents;
 use crate::reply_presentation::{
     append_prompt_if_missing, build_input_transcript, drop_echo_only_contents,
-    maybe_trim_echo_prefix, normalize_prompt, reconcile_polled_completion_prompt,
-    reconcile_trailing_completion_prompt, strip_trailing_prompt,
+    maybe_trim_echo_prefix, normalize_prompt, strip_trailing_prompt,
 };
 #[cfg(test)]
 use crate::reply_presentation::{
@@ -59,6 +58,7 @@ use crate::worker_supervisor::{
 };
 
 mod control_prefix;
+mod interrupt;
 mod output_state;
 mod pending_poll;
 mod write_dispatch;
@@ -1834,137 +1834,6 @@ impl WorkerManager {
         }
     }
 
-    fn interrupt_target_running(&mut self) -> Result<bool, WorkerError> {
-        match self.process.as_mut() {
-            Some(process) => process.is_running(),
-            None => Ok(false),
-        }
-    }
-
-    fn interrupt_files(
-        &mut self,
-        timeout: Duration,
-        deferred_sandbox_state_update: Option<SandboxStateUpdate>,
-        suppress_session_end_reset: bool,
-    ) -> Result<WorkerReply, WorkerError> {
-        crate::event_log::log(
-            "worker_interrupt_begin",
-            serde_json::json!({
-                "timeout_ms": timeout.as_millis(),
-            }),
-        );
-        let interrupt_drains_existing_completion =
-            self.pending_request || self.settled_pending_completion.is_some();
-        let should_interrupt_worker = self.interrupt_target_running()?;
-        if should_interrupt_worker {
-            let process = self
-                .process
-                .as_mut()
-                .expect("worker process should be available");
-            let interrupt_result = self.driver.interrupt(process);
-            if let Err(err) = interrupt_result {
-                self.reset()?;
-                crate::event_log::log(
-                    "worker_interrupt_error",
-                    serde_json::json!({
-                        "error": err.to_string(),
-                    }),
-                );
-                return Err(err);
-            }
-        }
-
-        if interrupt_drains_existing_completion {
-            let mut reply = self.poll_pending_output_files(timeout)?;
-            let prompt = match &reply.reply {
-                WorkerReply::Output { prompt, .. } => prompt.clone(),
-            };
-            let WorkerReply::Output { contents, .. } = &mut reply.reply;
-            reconcile_polled_completion_prompt(contents, prompt, self.backend);
-            let reply = self.finalize_reply(reply);
-            self.maybe_reset_after_session_end_with_options(
-                deferred_sandbox_state_update,
-                suppress_session_end_reset,
-                false,
-            )?;
-            return Ok(reply);
-        }
-
-        let mut timed_out = false;
-        let mut prompt: Option<String> = None;
-        if let Some(process) = self.process.as_ref()
-            && let Some(ipc) = process.ipc_connection()
-        {
-            let result = ipc.wait_for_prompt(timeout);
-            match result {
-                Ok(value) => {
-                    prompt = Some(value);
-                }
-                Err(IpcWaitError::Timeout) => {
-                    timed_out = true;
-                }
-                Err(IpcWaitError::SessionEnd) => {
-                    self.note_session_end(true);
-                }
-                Err(IpcWaitError::Disconnected) => {
-                    // IPC is optional for the R backend; fall back to prompt-as-output.
-                }
-                Err(IpcWaitError::Protocol(message)) => return Err(WorkerError::Protocol(message)),
-            }
-        }
-
-        let FormattedPendingOutput {
-            mut contents,
-            saw_stderr,
-        } = self.drain_formatted_output();
-        let is_error = saw_stderr;
-
-        if timed_out {
-            contents.push(timeout_status_content(timeout));
-        }
-
-        let session_end = self.session_end_seen;
-        let raw_prompt = if session_end || timed_out {
-            None
-        } else {
-            prompt.clone()
-        };
-        let resolved_prompt = normalize_prompt(raw_prompt.clone());
-        self.remember_prompt(raw_prompt);
-        if !session_end && !timed_out {
-            reconcile_trailing_completion_prompt(
-                &mut contents,
-                resolved_prompt.clone(),
-                self.backend,
-            );
-        }
-
-        let reply = WorkerReply::Output {
-            contents,
-            is_error,
-            error_code: timed_out.then_some(WorkerErrorCode::Timeout),
-            prompt: (!session_end).then_some(()).and(resolved_prompt),
-            prompt_variants: None,
-        };
-        crate::event_log::log(
-            "worker_interrupt_end",
-            serde_json::json!({
-                "timed_out": timed_out,
-                "session_end": session_end,
-            }),
-        );
-        let reply = self.finalize_reply(ReplyWithOffset {
-            reply,
-            end_offset: 0,
-        });
-        self.maybe_reset_after_session_end_with_options(
-            deferred_sandbox_state_update,
-            suppress_session_end_reset,
-            false,
-        )?;
-        Ok(reply)
-    }
-
     pub fn restart(&mut self, timeout: Duration) -> Result<WorkerReply, WorkerError> {
         match self.oversized_output {
             OversizedOutputMode::Files => self.restart_files(timeout),
@@ -2006,153 +1875,6 @@ impl WorkerManager {
         self.note_respawn_during_write();
         crate::event_log::log("worker_restart_end", serde_json::json!({"status": "ok"}));
         Ok(self.finalize_reply(reply))
-    }
-
-    fn interrupt_pager(
-        &mut self,
-        timeout: Duration,
-        deferred_sandbox_state_update: Option<SandboxStateUpdate>,
-        suppress_session_end_reset: bool,
-    ) -> Result<WorkerReply, WorkerError> {
-        crate::event_log::log(
-            "worker_interrupt_begin",
-            serde_json::json!({
-                "timeout_ms": timeout.as_millis(),
-            }),
-        );
-        let interrupt_drains_existing_completion =
-            self.pending_request || self.settled_pending_completion.is_some();
-        let should_interrupt_worker = self.interrupt_target_running()?;
-        if should_interrupt_worker {
-            let process = self
-                .process
-                .as_mut()
-                .expect("worker process should be available");
-            let interrupt_result = self.driver.interrupt(process);
-            if let Err(err) = interrupt_result {
-                self.reset()?;
-                crate::event_log::log(
-                    "worker_interrupt_error",
-                    serde_json::json!({
-                        "error": err.to_string(),
-                    }),
-                );
-                return Err(err);
-            }
-        }
-
-        let page_bytes = pager::resolve_page_bytes(None);
-        if interrupt_drains_existing_completion {
-            let mut reply = self.poll_pending_output_pager(timeout, page_bytes)?;
-            let pager_active = self.pager.is_active();
-            let prompt = match &reply.reply {
-                WorkerReply::Output { prompt, .. } => prompt.clone(),
-            };
-            let WorkerReply::Output { contents, .. } = &mut reply.reply;
-            if !pager_active {
-                reconcile_polled_completion_prompt(contents, prompt, self.backend);
-            }
-            let reply = self.finalize_reply(reply);
-            self.maybe_reset_after_session_end_with_options(
-                deferred_sandbox_state_update,
-                suppress_session_end_reset,
-                false,
-            )?;
-            return Ok(reply);
-        }
-
-        let mut timed_out = false;
-        let mut prompt: Option<String> = None;
-        if let Some(process) = self.process.as_ref()
-            && let Some(ipc) = process.ipc_connection()
-        {
-            let result = ipc.wait_for_prompt(timeout);
-            match result {
-                Ok(value) => {
-                    prompt = Some(value);
-                }
-                Err(IpcWaitError::Timeout) => {
-                    timed_out = true;
-                }
-                Err(IpcWaitError::SessionEnd) => {
-                    self.note_session_end(true);
-                }
-                Err(IpcWaitError::Disconnected) => {}
-                Err(IpcWaitError::Protocol(message)) => return Err(WorkerError::Protocol(message)),
-            }
-        }
-
-        let start_offset = self.output.current_offset().unwrap_or(0);
-        let mut end_offset = self.output.end_offset().unwrap_or(start_offset);
-        if end_offset < start_offset {
-            end_offset = start_offset;
-        }
-
-        let is_error = self
-            .output
-            .saw_stderr_in_range(start_offset.min(end_offset), end_offset);
-        let SnapshotWithImages {
-            mut contents,
-            pages_left,
-            buffer,
-            last_range,
-        } = snapshot_page_with_images(&self.output, end_offset, page_bytes);
-
-        if timed_out {
-            contents.push(timeout_status_content(timeout));
-        }
-
-        pager::maybe_activate_and_append_footer(
-            &mut self.pager,
-            &mut contents,
-            pages_left,
-            is_error,
-            buffer,
-            last_range,
-        );
-
-        let session_end = self.session_end_seen;
-        let raw_prompt = if session_end || timed_out {
-            None
-        } else {
-            prompt.clone()
-        };
-        let resolved_prompt = normalize_prompt(raw_prompt.clone());
-        self.remember_prompt(raw_prompt);
-        if self.pager.is_active() && !session_end {
-            self.pager_prompt = resolved_prompt.clone();
-        }
-        if !session_end && !timed_out && !self.pager.is_active() {
-            reconcile_trailing_completion_prompt(
-                &mut contents,
-                resolved_prompt.clone(),
-                self.backend,
-            );
-        }
-
-        let reply = WorkerReply::Output {
-            contents,
-            is_error,
-            error_code: timed_out.then_some(WorkerErrorCode::Timeout),
-            prompt: (!self.pager.is_active() && !session_end)
-                .then_some(())
-                .and(resolved_prompt),
-            prompt_variants: None,
-        };
-        crate::event_log::log(
-            "worker_interrupt_end",
-            serde_json::json!({
-                "timed_out": timed_out,
-                "session_end": session_end,
-            }),
-        );
-        let reply = self.finalize_reply(ReplyWithOffset { reply, end_offset });
-        self.maybe_reset_after_session_end_with_options(
-            deferred_sandbox_state_update,
-            suppress_session_end_reset,
-            false,
-        )?;
-        Ok(reply)
     }
 
     fn restart_pager(&mut self, timeout: Duration) -> Result<WorkerReply, WorkerError> {
